Creating project & preparing server class
