+++ StandaloneServer/Program.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace CSM.StandaloneServer
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+        }
+    }
+}
